+++ src/tcf/component/Banner.tsx
